+++ src/Items/Item.jsx
@@ -5,14 +5,14 @@ import specialAssign from '../utils/special-assign'
 
 const checkedProps = {
   component: PropTypes.any,
-  id:        PropTypes.string,
-  role:      PropTypes.string,
-  index:     PropTypes.number,
-  text:      PropTypes.string,
-  value:     PropTypes.any,
-  onFocus:   PropTypes.func,
-  onSelect:  PropTypes.func,
-  children:  PropTypes.oneOfType([PropTypes.func, PropTypes.node])
+  id: PropTypes.string,
+  role: PropTypes.string,
+  index: PropTypes.number,
+  text: PropTypes.string,
+  value: PropTypes.any,
+  onFocus: PropTypes.func,
+  onSelect: PropTypes.func,
+  children: PropTypes.oneOfType([PropTypes.func, PropTypes.node])
 }
 const noop = () => null
 
@@ -25,50 +25,51 @@ class Item extends Component {
 
   static defaultProps = {
     component: 'div',
-    role:      'menuitem',
-    onFocus:   noop,
-    onSelect:  noop
+    role: 'menuitem',
+    onFocus: noop,
+    onSelect: noop
   }
 
   _id = this.props.id || uuid()
 
   componentDidMount() {
-    const { itemList: { focusGroup } } = this.context
-    const { index, text, value } = this.props
+    const { focusGroup } = this.context.itemList
+    const { index, value, label } = this.props
+    const node = findDOMNode(this)
 
     this._member = {
-      id:   this._id,
-      node: findDOMNode(this),
+      id: this._id,
+      node,
       index,
-      text,
-      value
+      value,
+      label: label || node.innerHTML
     }
 
-    // add this item as a member
     focusGroup.addMember(this._member)
 
-    // listen for respective focus group events
     focusGroup.on('focus', this._handleMemberFocus)
     focusGroup.on('select', this._handleMemberSelect)
 
-    // activate focus group if this was the first member added
     if (focusGroup.getMembers().length === 1) {
       focusGroup.activate()
     }
   }
 
   componentWillUnmount() {
-    const { itemList: { focusGroup } } = this.context
+    const { focusGroup } = this.context.itemList
 
     focusGroup.removeMember(this._member)
 
     focusGroup.off('focus', this._handleMemberFocus)
     focusGroup.off('select', this._handleMemberSelect)
 
-    // deactivate focus group if this was the last member removed
     if (focusGroup.getMembers().length <= 0) {
       focusGroup.deactivate()
     }
+  }
+
+  getMember() {
+    return this._member
   }
 
   _handleMemberFocus = (member, e) => {
@@ -83,18 +84,12 @@ class Item extends Component {
     }
   }
 
-  _handleClick = (e) => {
-    this.props.onSelect(this._member, e)
-    this.context.itemList.onItemSelection(this._member, e)
-  }
-
   render() {
     const { component, role, children } = this.props
     const props = specialAssign({
       role,
       id: this._id,
-      tabIndex: -1,
-      onClick: this._handleClick
+      tabIndex: -1
     }, this.props, checkedProps)
 
     if (typeof children === 'function') {