+++ src/utils/get-scroll-parent.js
@@ -1,0 +1,10 @@
+export default function getScrollParent(node) {
+  if (node === null) {
+    return null
+  }
+  if (node.scrollHeight > node.clientHeight) {
+    return node
+  } else {
+    return getScrollParent(node.parentNode)
+  }
+}